requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
motor>=3.6.0
pymongo>=4.10.1
pydantic>=2.9.0
pydantic-settings>=2.6.0
httpx>=0.27.0
python-dotenv>=1.0.0
python-multipart>=0.0.12
apscheduler>=3.10.4
python-socketio>=5.11.0
websockets>=12.0
